Configure Document and Window Objects on the server

+++ src/app/services/local-storage.service.ts
@@ -1,18 +1,28 @@
-import { Injectable } from '@angular/core';
+import { Injectable, PLATFORM_ID, Inject } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable()
 export class LocalStorageService {
+  constructor(@Inject(PLATFORM_ID) private platformId: object) { }
+
   // pass a value you want to get from the localStorage as the string `value`.
   // if the value is not found, pass a fallback. It can be anything.
   get(key: string, fallback: any) {
-    const value = localStorage.getItem(key);
-
-    return value ? value : fallback;
+    if (isPlatformBrowser(this.platformId)) {
+      const value = localStorage.getItem(key);
+      return value ? value : fallback;
+    }
   }
   set(key: string, value: any) {
-    localStorage.setItem(key, value);
+
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem(key, value);
+    }
   }
   clear() {
-    localStorage.clear();
+
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
   }
 }
